05_tools/pyproject.toml
[project]
name = "05-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.0.19",
    "tavily-python>=0.7.2",
]
